ai-lightning/server/requirements.txt
Flask==3.0.0
Flask-SocketIO==5.3.5
Flask-JWT-Extended==4.4.4
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.9
redis==5.0.1
pyln-client==0.12.1
gunicorn==21.2.0
python-dotenv==1.0.0
httpx==0.25.2
Werkzeug==3.0.1
python-socketio==5.10.0
eventlet==0.34.2
click==8.1.7
pytest==7.4.3
pytest-flask==1.3.0